drivers: fbdev: dpu_everest: change idma_type usage.
idma_type of S3CFB_WIN_CONFIG parameter will be changed to channel number
in the future.

For preparing this change, idma_type is used as DPP channel number
in driver side.

However, it is remained to use idma_type for BTS and protection code.
Because it is needed to modify another driver. So, it will be changed later.

Change-Id: Ida1d2fc40664eeafa1f76901044ec6cfaeedf79d
Signed-off-by: Jiun Yu <jiun.yu@samsung.com>